consolidating e2e in one file because of a bug in chromeless.end()
