+++ test/e2e/collective.loggedout.test.js
@@ -8,7 +8,7 @@ const WEBSITE_URL = "https://staging.opencollective.com";
 const screenshotsDirectory = (process.env.CIRCLE_ARTIFACTS) ? process.env.CIRCLE_ARTIFACTS : '/tmp';
 console.log(">>> screenshotsDirectory", screenshotsDirectory);
 
-describe("collective page logged out", () => {
+describe("logged out", () => {
   let chromeless;
 
   before((done) => {
@@ -16,8 +16,8 @@ describe("collective page logged out", () => {
     done();
   })
 
-  after(() => chromeless.end());
-
+  after(async () => await chromeless.end());
+  
   it("load collective page", async function() {
     
     this.timeout(20000);
@@ -70,4 +70,51 @@ describe("collective page logged out", () => {
     console.log(">>> url", url);
     expect(url).to.contain(`${WEBSITE_URL}/webpack/expenses/new`);
   });
+
+  it("makes a one time donation", async function() {
+    
+    this.timeout(25000);
+
+    const run = async () => {
+
+      const email = `testuser+${Math.round(Math.random()*1000000)}@gmail.com`;
+
+      const screenshot = await chromeless
+        .goto(`${WEBSITE_URL}/webpack/donate`)
+        .type(email, "input[name='email']")
+        .type("Xavier", "input[name='firstName']")
+        .type("Damman", "input[name='lastName']")
+        .type("https://xdamman.com", "input[name='website']")
+        .type("xdamman", "input[name='twitterHandle']")
+        .type("short bio", "input[name='description']")
+        .type("4242424242424242", "input[name='CCnumber']")
+        .type("Full Name", "input[name='CCname']")
+        .type("11/22", "input[name='CCexpiry']")
+        .type("111", "input[name='CCcvc']")
+        .click(".presetBtn")
+        .type("Public message", "textarea[name='publicMessage']")
+        .click('.submit button')
+        .wait('.UserCollectivePage', 10000)
+        .screenshot({ filePath: path.join(screenshotsDirectory, 'onetime_donation.png')});
+
+      console.log(">>> screenshot", screenshot);
+      const url = await chromeless.evaluate(() => window.location.href)
+      console.log(">>> url", url);
+      expect(url).to.contain(`${WEBSITE_URL}/xdamman`);
+      expect(url).to.contain(`?status=orderCreated&CollectiveId=302`);
+      const thankyou = await chromeless.exists('p.thankyou');
+      expect(thankyou).to.be.true;
+      const messageContent = await chromeless.evaluate(() => document.querySelector('.message').innerText);
+      expect(messageContent).to.contain('webpack');
+    }
+
+    try {
+      await run();
+    } catch (e) {
+      // Sadly this doesn't work yet: https://github.com/graphcool/chromeless/issues/279
+      const screenshot = await chromeless.screenshot({ filePath: path.join(screenshotsDirectory, 'collective_page.png')});
+      console.error(">>> error: ", e);
+      console.log(">>> screenshot", screenshot);
+    }
+  })
 });
